very noisy debugging for branch

+++ sip_router/msg_parser.h
@@ -1,5 +1,5 @@
 /*
- * $Id: msg_parser.h,v 1.21 2001/12/02 18:28:24 jku Exp $
+ * $Id: msg_parser.h,v 1.22 2001/12/02 19:21:49 jku Exp $
  */
 
 #ifndef msg_parser_h
@@ -177,7 +177,7 @@ struct sip_msg{
 
 	/* str add_to_branch; */ /* whatever whoever want to append to branch comes here */
 	char add_to_branch_s[MAX_BRANCH_PARAM_LEN];
-	short add_to_branch_len;
+	int add_to_branch_len;
 
 	
 };
